TargetDummies, AudioSystem, Particles!
Added a sandbox level with two target dummies.
Added an AudioManager that works as a singleton to plau soundeffects.,
Added AudioSources to spells and characters so it is possible to spawn sound effects when casting, spawning and hitting targets.
Added particle effects to spawn at every spell (needed for sound effects above)